+++ src/main/java/com/vision/project/repositories/base/ChatRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import java.util.List;
 
-public interface ChatRepository extends JpaRepository<Chat, Integer> {
+public interface ChatRepository extends JpaRepository<Chat, Long> {
     @Query(value="from Chat where first_user = :user or second_user = :user order by id")
-    List<Chat> findUserChats(@Param("user") int id, Pageable pageable);
+    List<Chat> findUserChats(@Param("user") long id, Pageable pageable);
 }
