+++ src/main/java/com/vision/project/repositories/base/DishRepository.java
@@ -3,5 +3,5 @@ package com.vision.project.repositories.base;
 import com.vision.project.models.Dish;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DishRepository extends CrudRepository<Dish, Integer> {
+public interface DishRepository extends CrudRepository<Dish, Long> {
 }
