+++ src/main/java/com/vision/project/repositories/base/MenuRepository.java
@@ -4,7 +4,6 @@ import com.vision.project.models.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface MenuRepository extends JpaRepository<Menu, Integer> {
-
-    List<Menu> findByRestaurant(int restaurant);
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+    List<Menu> findByRestaurant(long restaurant);
 }
